reworked how async was being passed and now download component is much simpler
